requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
Pillow==10.0.1
drf-yasg==1.21.5
drf-spectacular==0.27.1
dj-database-url==2.1.0
gunicorn==21.2.0
